chore: move `docker_build` logic to lib (#211)
![image](https://github.com/user-attachments/assets/6d297e22-2dd4-48d5-b84d-61e2138ac9b4)
