+++ cargo-near-build/Cargo.toml
@@ -20,3 +20,29 @@ near-abi = { version = "0.4.0", features = ["__chunked-entries"] }
 zstd = "0.13"
 schemars = "0.8"
 rustc_version = "0.4"
+url = { version = "2.5.0", features = ["serde"], optional = true }
+serde = { version = "1.0.197", optional = true }
+git2 = { version = "0.19", optional = true }
+home = { version = "0.5.9", optional = true }
+pathdiff = { version = "0.2.1", features = ["camino"], optional = true }
+unix_path = { version = "1.0.1", optional = true }
+tempfile = { version = "3.10.1", optional = true }
+
+[target.'cfg(target_os = "linux")'.dependencies]
+nix = { version = "0.29.0", features = ["user", "process"], optional = true }
+
+[package.metadata.docs.rs]
+features = ["build_script"]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[features]
+default = []
+build_script = []
+cli_exports = [] 
+docker = [
+  "cli_exports",
+  "dep:url", "dep:serde", "dep:git2", 
+  "dep:home", "dep:pathdiff", "dep:unix_path", 
+  "dep:tempfile", "dep:nix"
+]
+
