+++ cargo-near-build/src/near/abi/generate/mod.rs
@@ -1,9 +1,10 @@
 use std::collections::HashMap;
 
+use crate::types::near::build::input::ColorPreference;
 use crate::{
     cargo_native::{self, Dylib},
     env_keys, pretty_print,
-    types::{cargo::metadata::CrateMetadata, color_preference::ColorPreference},
+    types::cargo::metadata::CrateMetadata,
 };
 use eyre::ContextCompat;
