+++ cargo-near-build/src/near/docker_build/git_checks/mod.rs
@@ -1,0 +1,2 @@
+pub mod dirty;
+pub mod pushed_to_remote;
