+++ cargo-near-build/src/types/near/docker_build/cloned_repo.rs
@@ -1,14 +1,21 @@
+use super::BuildContext;
 use std::marker::PhantomData;
+use std::time::Duration;
 
-use cargo_near_build::{
-    camino, ArtifactMessages, BuildArtifact, CrateMetadata, ManifestPath, MANIFEST_FILE_NAME,
-};
-use cargo_near_build::{pretty_print, VersionMismatch};
+use crate::pretty_print;
+use crate::types::cargo::manifest_path::{ManifestPath, MANIFEST_FILE_NAME};
+use crate::types::cargo::metadata::CrateMetadata;
+use crate::types::near::build::output::version_mismatch::VersionMismatch;
+use crate::types::near::build::side_effects::ArtifactMessages;
+use crate::types::near::docker_build::WARN_BECOMES_ERR;
+use crate::{camino, BuildArtifact};
 use colored::Colorize;
 
 use super::crate_in_repo;
 
-pub(super) struct ClonedRepo {
+const ERR_NO_LOCKED_DEPLOY: &str = "`--no-locked` flag is forbidden for deploy with docker.";
+
+pub struct ClonedRepo {
     pub initial_crate_in_repo: crate_in_repo::Crate,
     #[allow(unused)]
     pub tmp_repo_dir: tempfile::TempDir,
@@ -17,10 +24,31 @@ pub(super) struct ClonedRepo {
 }
 
 impl ClonedRepo {
-    pub(super) fn git_clone(
+    pub fn check_locked_then_clone(
         crate_in_repo: crate_in_repo::Crate,
         no_locked: bool,
-    ) -> color_eyre::eyre::Result<Self> {
+        context: BuildContext,
+    ) -> eyre::Result<Self> {
+        match (no_locked, context) {
+            (false, _) => {}
+            (true, BuildContext::Build) => {
+                no_locked_warn_pause(true);
+                println!();
+                println!("{}", WARN_BECOMES_ERR.red(),);
+                std::thread::sleep(Duration::new(5, 0));
+            }
+            (true, BuildContext::Deploy) => {
+                println!(
+                    "{}",
+                    "Check in Cargo.lock for contract being built into source control.".yellow()
+                );
+                return Err(eyre::eyre!(ERR_NO_LOCKED_DEPLOY));
+            }
+        }
+        Self::git_clone(crate_in_repo, no_locked)
+    }
+
+    fn git_clone(crate_in_repo: crate_in_repo::Crate, no_locked: bool) -> eyre::Result<Self> {
         let tmp_repo_dir = tempfile::tempdir()?;
         let tmp_repo_path = tmp_repo_dir.path().to_path_buf();
         let tmp_repo =
@@ -41,7 +69,7 @@ impl ClonedRepo {
             };
             CrateMetadata::collect(ManifestPath::try_from(cargo_toml_path)?, no_locked).map_err(|err| {
             if !no_locked && err.to_string().contains("Cargo.lock is absent") {
-                super::no_locked_warn_pause(false);
+                no_locked_warn_pause(false);
                 println!();
                 println!("{}", "Cargo.lock check was performed against git version of code.".cyan());
                 println!("{}", "Don't forget to check in Cargo.lock into source code for deploy if it's git-ignored...".cyan());
@@ -62,13 +90,13 @@ impl ClonedRepo {
         })
     }
 
-    pub(super) fn crate_metadata(&self) -> &CrateMetadata {
+    pub fn crate_metadata(&self) -> &CrateMetadata {
         &self.tmp_crate_metadata
     }
-    pub(super) fn copy_artifact(
+    pub fn copy_artifact(
         self,
-        cli_override: Option<crate::types::utf8_path_buf::Utf8PathBuf>,
-    ) -> color_eyre::eyre::Result<BuildArtifact> {
+        cli_override: Option<camino::Utf8PathBuf>,
+    ) -> eyre::Result<BuildArtifact> {
         let tmp_out_dir = self.tmp_crate_metadata.resolve_output_dir(None)?;
 
         let destination_crate_metadata = {
@@ -91,7 +119,7 @@ fn copy(
     tmp_out_dir: camino::Utf8PathBuf,
     tmp_crate_metadata: CrateMetadata,
     mut destination_dir: camino::Utf8PathBuf,
-) -> color_eyre::eyre::Result<BuildArtifact> {
+) -> eyre::Result<BuildArtifact> {
     println!(
         " {} {}",
         "artifact search location in temporary build site:".green(),
@@ -103,7 +131,7 @@ fn copy(
     let in_wasm_path = tmp_out_dir.join(filename.clone());
 
     if !in_wasm_path.exists() {
-        return Err(color_eyre::eyre::eyre!(
+        return Err(eyre::eyre!(
             "Temporary build site result wasm file not found: `{:?}`.",
             in_wasm_path
         ));
@@ -130,4 +158,20 @@ fn copy(
     messages.pretty_print();
 
     Ok(result)
+}
+
+fn no_locked_warn_pause(warning_red: bool) {
+    println!();
+    let warning = if warning_red {
+        format!("{}", "WARNING: ".red())
+    } else {
+        "".to_string()
+    };
+    println!(
+        "{}{}",
+        warning,
+        "Please mind that `--no-locked` flag is allowed in Docker builds, but:".cyan()
+    );
+    println!("{}", "  - such builds are not reproducible due to potential update of dependencies and compiled `wasm` mismatch as the result.".yellow());
+    std::thread::sleep(Duration::new(12, 0));
 }
