+++ cargo-near-build/src/types/near/mod.rs
@@ -1,3 +1,6 @@
 pub mod abi;
 pub mod build;
+#[cfg(feature = "build_script")]
 pub mod build_extended;
+#[cfg(feature = "docker")]
+pub mod docker_build;
