+++ cargo-near/Cargo.toml
@@ -23,13 +23,12 @@ license = false
 eula = false
 
 [dependencies]
-cargo-near-build = { version = "0.1.0", path = "../cargo-near-build" }
+cargo-near-build = { version = "0.1.0", path = "../cargo-near-build", features = ["docker"] }
 clap = { version = "4.0.18", features = ["derive", "env"] }
 colored = "2.0"
 env_logger = "0.11.5"
 log = "0.4"
 serde = "1.0.197"
-serde_json = "1.0"
 color-eyre = "0.6"
 inquire = "0.7"
 strum = { version = "0.24", features = ["derive"] }
@@ -41,12 +40,6 @@ shell-words = "1.0.0"
 interactive-clap = "0.3"
 interactive-clap-derive = "0.3"
 near-cli-rs = { version = "0.14.0", default-features = false }
-tempfile = "3.10.1"
-git2 = "0.19"
-home = "0.5.9"
-pathdiff = { version = "0.2.1", features = ["camino"] }
-unix_path = "1.0.1"
-url = { version = "2.5.0", features = ["serde"] }
 reqwest = "0.12.5"
 tracing = "0.1.40"
 tokio = { version = "1.0", default-features = false, features = [
@@ -56,8 +49,6 @@ tokio = { version = "1.0", default-features = false, features = [
 [dev-dependencies]
 tracing-test = "0.2.5"
 
-[target.'cfg(target_os = "linux")'.dependencies]
-nix = { version = "0.29.0", features = ["user", "process"] }
 
 [features]
 default = ["ledger"]
