--- cargo-near/src/commands/build_command/docker/git_checks.rs
@@ -1,2 +1,0 @@
-pub(super) mod dirty;
-pub(super) mod pushed_to_remote;
