+++ cargo-near/src/types/mod.rs
@@ -1,3 +1,2 @@
 pub mod color_preference_cli;
-pub mod source_id;
 pub mod utf8_path_buf;
