+++ integration-tests/Cargo.toml
@@ -6,13 +6,13 @@ publish = false
 
 [dependencies]
 const_format = "0.2"
-cargo-near-build = { path = "../cargo-near-build" }
+cargo-near-build = { version = "0.1.0", path = "../cargo-near-build" }
 
 [dev-dependencies]
 borsh = { version = "1.0.0", features = ["derive", "unstable__schema"] }
 camino = "1.1.1"
 cargo-near = { path = "../cargo-near" }
-cargo-near-build = { path = "../cargo-near-build" }
+cargo-near-build = { version = "0.1.0", path = "../cargo-near-build" }
 color-eyre = "0.6"
 function_name = "0.3"
 git2 = "0.19"
